+++ src/QueryCat.Backend.Core/Execution/IExecutionThread.cs
@@ -36,16 +36,6 @@ public interface IExecutionThread : IDisposable
     IExecutionStack Stack { get; }
 
     /// <summary>
-    /// The event is fired before variable resolving.
-    /// </summary>
-    event EventHandler<ResolveVariableEventArgs>? VariableResolving;
-
-    /// <summary>
-    /// The event is fired after variable resolve.
-    /// </summary>
-    event EventHandler<ResolveVariableEventArgs>? VariableResolved;
-
-    /// <summary>
     /// Selector to resolve object expressions.
     /// </summary>
     IObjectSelector ObjectSelector { get; }
@@ -75,15 +65,6 @@ public interface IExecutionThread : IDisposable
         string query,
         IDictionary<string, VariantValue>? parameters = null,
         CancellationToken cancellationToken = default);
-
-    /// <summary>
-    /// Try to get variable value from top scope to the root recursively.
-    /// </summary>
-    /// <param name="name">Variable name.</param>
-    /// <param name="value">Variable value.</param>
-    /// <param name="scope">Scope instance. Top scope is used by default.</param>
-    /// <returns>True if variable with the specified name is found, false otherwise.</returns>
-    bool TryGetVariable(string name, out VariantValue value, IExecutionScope? scope = null);
 
     /// <summary>
     /// Get completions for the incomplete query.