+++ src/QueryCat.Backend/ExecutionThreadBootstrapper.cs
@@ -49,6 +49,8 @@ public sealed class ExecutionThreadBootstrapper(ExecutionOptions? options = null
 
     private bool _addStandardCompletions;
 
+    private Func<IExecutionScope?, IExecutionScope>? _executionScopeFactory;
+
     /// <summary>
     /// Use the custom config storage.
     /// </summary>
@@ -182,6 +184,12 @@ public sealed class ExecutionThreadBootstrapper(ExecutionOptions? options = null
         return this;
     }
 
+    public ExecutionThreadBootstrapper WithExecutionScopeFactory(Func<IExecutionScope?, IExecutionScope> factory)
+    {
+        _executionScopeFactory = factory;
+        return this;
+    }
+
     /// <summary>
     /// Create the instance of execution thread.
     /// </summary>
@@ -230,7 +238,8 @@ public sealed class ExecutionThreadBootstrapper(ExecutionOptions? options = null
             objectSelector: _objectSelector,
             configStorage: _inputConfigStorage,
             astBuilder: astBuilder,
-            completionSource: completionSource
+            completionSource: completionSource,
+            executionScopeFactory: _executionScopeFactory
         );
         thread.Tag = _tag;